backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.3
tzdata==2024.1
psycopg[binary]==3.2.1
python-dotenv
bcrypt
djangorestframework
drf-yasg